# backend/main-fiber/docs/swagger.yaml
basePath: /api
definitions:
...
  api.Position:
    properties:
      constraints:
        properties:
          exclusive_per_org:
            description: 1 per org?
            type: boolean
          max_slots:
            description: capacity (nil = unlimited)
            type: integer
          term_days:
            description: optional expiry window
            type: integer
        type: object
      created_at:
        type: string
      display:
        additionalProperties:
          type: string
        description: '{ "en": "Head" }'
        type: object
      key:
        description: stable id (e.g., "head")
        type: string
      rank:
        description: sort order
        type: integer
      scope:
        properties:
          inherit:
            description: usable by descendants?
            type: boolean
          org_path:
            description: where defined ("/" for global)
            type: string
        type: object
      status:
        description: '"active" | "deprecated"'
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: This is my API
  title: My API
  version: "1.0"
paths:
...
  /positions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.Position'
            type: array
      summary: List positions
      tags:
      - positions
    post:
      consumes:
      - application/json
      parameters:
      - description: Position
        in: body
        name: position
        required: true
        schema:
          $ref: '#/definitions/api.Position'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.Position'
      summary: Create position
      tags:
      - positions
  /positions/{key}:
    delete:
      parameters:
      - description: Position key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Deprecate position (soft delete)
      tags:
      - positions
    get:
      parameters:
      - description: Position key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Position'
      summary: Get a position by key
      tags:
      - positions
    put:
      consumes:
      - application/json
      parameters:
      - description: Position key
        in: path
        name: key
        required: true
        type: string
      - description: Position (partial)
        in: body
        name: position
        required: true
        schema:
          $ref: '#/definitions/api.Position'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Position'
      summary: Update position
      tags:
      - positions
...
swagger: "2.0"